feat: add debug infos and move ResetAppBtn to correct folder
