+++ package.json
@@ -24,6 +24,7 @@
     "expo-application": "~6.0.2",
     "expo-build-properties": "^0.13.2",
     "expo-check-installed-apps": "^0.2.9",
+    "expo-clipboard": "~7.0.1",
     "expo-constants": "~17.0.2",
     "expo-dev-client": "^5.0.12",
     "expo-device": "~7.0.2",
